Add List.cpp
Design a class for the following scenario:
An item list contains item code, name, rate, and quantity for several
items.
Whenever a new item is added in the list uniqueness of item code is to
be checked.
Time to time rate of the items may change.
Whenever an item is issued or received existence of the item is checked
and quantity is updated.
In case of issue, availability of quantity is also to be checked.
User may also like to know price/quantity available for an item.