Add Pascal.cpp
Pascal’s Triangle has the interesting property that every entry is the
sum of the two entries above it, except along the left and right edges,
where the values are always 1. Consider, for example, the highlighted
entry in the following display of Pascal’s.
Write a simple test program to demonstrate that your combinations
function works. You may also write a program that uses c(n, k) to
display the first ten rows of Pascal’s Triangle.