Add array.cpp
Design an ARRAY class with the following features:
a. Array object may be declared for a specific size and a value for
initialising all the elements. If this it is to be assumed as a 0.
b. An array object may be declared and initialised with another object.
c. An array object may be declared and initialised with another array
(not the object, standard array as in C language). Let a and b are two
objects:
i. a+b will add corresponding elements.
ii. a=b will do the assignment.
iii. a[I] will return the ith element of the object.
iv. a*5 or 5*a will multiply the element with 5.