Add prog. to list items in a dir(sorted)
A program that lists all of the files and subdirectories in a specified
directory, along with their sizes and modification dates. By default,
the output is by name.
If invoked with the -s option, however, output is sorted by size, from
largest to smallest.
If invoked with the -d option, output is sorted by date, from most
recent to least.