Add package related to Student
Design a STUDENT class to store roll, name, course, admission date and
marks in 5 subjects taken from user. Create an array of STUDENT
objects. Provide methods corresponding to admission date and receiving
marks, preparing mark sheet. Support must be there to show the number
of students who have taken admission.

Inherit Student class and override the input method so as to input the
department of each student. Search and display a sorted list of
students of one department or students based on scoring criteria.
Create an arraylist of students and remove a student based on certain
criterion and then call gc() and check for free memory.