Add solution for QuoteOfTheDay (#4; Package)
Create a class diagram and Java code for the following system and
scenario, taking into account the possibility of future extensions. The
system is a command line utility that prints a short 'quote of the day'
on the user's terminal when run. To begin with the quote is selected
randomly from a set of hard- coded strings within the program itself.
Examples:
> javac QuoteOfTheDay
God helps them that help themselves. -- Benjamin Franklin