Add solution for Hospital (#6; Package)
Basic solution to :

Design and create a hospital information system with the following
scenarios.
• Register a new patient.
• Each patient is assigned to one doctor, but a doctor can have any
number of patients. Patients
check in to the hospital and assigned a doctor if they don't already
have one.
• While in the hospital, doctors record various observations about each
patient at various times.
Examples of observations are blood pressure and temperature. Record
test results for a patient.
• The hospital keeps track of all the observations for a given patient
until they check out of the
hospital. Obtain all of a patient’s information given the social
security number.