Add solution for sorted output of SubDirectoryAndFiles  (#7; Package)
Write a program that lists all of the files and subdirectories in a
specified directory, along with their sizes and modification dates.
By default, the output should be sorted by name.
If invoked with the -s option, however, output should be sorted by
size, from largest to smallest.
If invoked with the -d option, output should be sorted by date, from
most recent to least.